rust: Remove absolute paths from dep-info files
We use set an absolute path for OUT_DIR as otherwise `include!` macros
search for generated sources in a path relative to the source file.

However this means that our dep-info files contain absolute paths. This
can cause Soong to not find a path if the dep-info file comes from a
remote build environment with a different $PWD, resulting in Soong
unnecessarily rebuilding the module because it believes the inputs have
been updated.

This CL adds a post-processing step for our dep-info file to the
rustc_wrapper script to strip out $PWD if it's present in any of the
paths.

Bug: 407072297
Test: m libprotobuf # check deps file for absolute paths.
Change-Id: I806223483726d1359ced2d98e80e4fc84010b79e