Allow overriding -Werror=bool-operation.
This warning also turns on -Werror=bitwise-instead-of-logical,
which does not always indicate a bug and is in fact triggered
by the correctly working code in external/llvm. Therefore
it is not appropriate to always treat it as an error.

Move the warning to commonGlobalCflags, where it can be
overridden from Android.bp, and remove the global disable
of -Wno-bitwise-instead-of-logical from mandatory 3P flags.

Bug: 428296962
Test: presubmit
Flag: EXEMPT warning flag changes, no impact on runtime behavior
Change-Id: I6e59f38140c3afdcaa7fee24a5b3e0fdb6d6db13