Diasble ImageMutator on genrules
genrule modules have to implement ImageInterface so that cc_genrule
modules can extend them with image mutator support.  Add a new
ImageMutatorSupported method to ImageInterface that allows it to
implement ImageInterface but still  not have any variants created
and always forward incoming dependencies to the empty variation.
This avoids needing to rely on the single variant fallback for this
case.

Bug: 424439549
Test: builds
Change-Id: I8e46c166089e8477039e388d887f1060bd5afcdf