Remove duplicate flags for -Wimplicit-int-float-conversion.
This warning is enabled, enabled again, then disabled in this
section. This is rather confusing.

This warning is not included in -Wall or -Wextra, but several
modules enable it through -Wconversion and trigger the warning,
so for now we still need to explicitly disable it.

Bug: 145211066
Bug: 296606895
Test: presubmit
Flag: EXEMPT warning flag cleanup
Change-Id: Ic596827ff80cb3a1281bccb740046d53d7642b39