Reorder and document -Werror flags.
The -Werror flags in noOverrideGlobalCflags are interspersed
with temporary warning disables, which makes it hard to see
what is enabled. Put them all at the top of the section and
add a description for each warning. Similarly, add comments
for each -Werror flag in deviceGlobalCflags.

Bug: 296606895
Test: presubmit
Flag: EXEMPT no-op warning flag cleanup
Change-Id: Ie75851e0173defedd219850a36bca6e2e7e965da