Support pvmfw in vbmeta images
pvmfw avb hashtree is included in vbmeta_system.img, but this is
currently missing from the soong autogenerated vbmeta file. This CL adds
this support to Soong.

To support this, vbmetaPartitionInfo of bootimg has been updated to use
`Partition_name` as the name for vbmeta. This matches the description of this
property. To prevent breaking the behavior for the autogenerarted boot,
init_boot, ... boot images, this property has been made explicit in
fsgen.

Bug: 424812534
Bug: 419860251

Test: verified that vbmeta_system.img in target_files.zip is identical
for husky

Change-Id: If902a4e9060021fec712c2ee0f36a51165030d46