Split shouldGeneratePhonyTargets out of shouldSkipAndroidMkProcessing
generateModuleTarget was conflating shouldSkipAndroidMkProcessing
with NamespaceExportedToMake, causing it to no longer emit phony
rules for unexported namespaces.  Split shouldGeneratePhonyTargets
out of shouldSkipAndroidMkProcessing and don't call generatePhonyTargets
at all when it isn't needed, and then use ExportedToMake inside
generatePhonyTargets to determine whether or not to create the base
phony target or only the namespaced ones.

Bug: 417632029
Test: aninja -t query apache-xml-soong
Test: aninja -t query SystemUI-soong
Change-Id: Iedf9d1fda484eea6cbddf1dc6101996865e15183