Change the module phony behavior to only build installed or output files
Running `m SystemUI` in a Soong-only build is building more than it
does in a Soong+Make build because Make is not building Soong's extra
checkbuild files due to not making LOCAL_CHECKED_MODULE a dependency
of the module phony.  This slows down inner loop development by building
unnecessary files like the turbine jars.

This patch modifies the Soong module phony targets to build only
the installed files if they exist, or the output files if there are
no installed files, or falling back to the checkbuild files if there
are no output files.

All of the files can be built by specifiying `m SystemUI-checkbuild` if
desired.

Bug: 417632029
Test: aninja -t query SystemUI-soong
Change-Id: I6f26321c79fae4fa2021e5e6c6a169f40e393f49