Fix duplicated vintf inter partition installs
This is a followup fix to the inter partition vintf installs implemented
in ag/34413806. ag/34413806 removed all cross partition vintfs including
vintfs of recovery binaries, which are expected to be installed in
vendor_ramdisk.

This CL adds a boolean to interPartitionDepTag to determine which
cross partition vintfs should be skipped.

Test: m nothing --no-skip-soong-tests
Bug: 430512479
Change-Id: I6ab251cea5287cd416844d2ba2122b7e1158de1d