Implement prebuilt system image replacement
prebuilt_module_name will replace the output image ($OUT/system.img)
while retaining all other build rules. Also modules in the partition
won't be built on "m droid" or "m dist", meaning that $OUT and target
files zip won't have outputs of such modules (but still can be built
invidivually by "m {module-name}"). Instead, files inside the prebuilt
system.img will be exported to target files zip.

The ultimate goal is to support the prebuilt system image. But that's a
big project, and we can't implement a perfect solution from the
beginning. Instead, this is the first step of the prebuilt system image.
The motivation of this change is to support other partitions depending
on the prebuilt partitions. For example, system_other.img contains
outputs of dexpreopt from /system, /system_ext, and /product partitions.
If we were to replace the entire system image module, that would cut
dependencies to all other build rules like apps and it will be extremely
tricky to get required artifacts for system_other.img because we now
have to unwrap the prebuilt image.

Bug: 419170198
Test: build
Change-Id: I638a1d20397053fadc185b0954a23a3ef29884b0