Handle fsgen partitions of override modules that use `//` syntax for base
Some override modules use //$some_dir:$module syntax to specify the base
module instead of a namespace qualified syntax.
`updatePartitionsOfOverrideModules` uses a fully qualified name to get
the partition of the base app. If the first component of `base` is not
exported via PRODUCT_SOONG_NAMESPACES, it falls back to the second
component in root namespace.

This heuristic runs into some issues when the first component is not a
valid namespace, but is specified in PRODUCT_SOONG_NAMESPACES.

This CL introduces a temporary solution to always use the second
component of `base`. To prevent incorrectly guessing the partition of
the base app, a check has been added to assert that there is no
partition skew of base app across namespaces.

Bug: 430646480
Test: diff test for shiba
Change-Id: Ieeca11cae6d180188995c918902026a1e48a7891