rust: bindgen: Block `max_align_t` by default
`max_align_t` in Clang is defined using a `long double` on i686.
Unfortunately, that type is 12 bytes in size. Rust has no ability to
represent a 12 byte immediate.

Luckily, `max_align_t` is essentially never used directly, only for its
properties in macros and expansions, and so it should not actually
appear in bindings for actual functions.

Bug: 432074924
Test: TH
Change-Id: Id65633469ef696a9b377485cfe241468e1e79a89