Make namespace imports strict
Previously, namespace imports would search for a namespace starting
at the specified path but then walking up the parent folders. It would
even find the root namespace if there were no other namespaces.

This is because findNamespace(), which walks up, was storing its results
into namespacesByDir, which is used by namespaceAt(), which is not
supposed to walk up.

Create a dedicated cache for findNamespace() instead.

Bug: 432305765
Test: m nothing --no-skip-soong-tests
Change-Id: Iee098d91c37b1e20007a411b66b80da628294e38