Add a script to find the product mk file from product name
The script finds the mk file that defines the product, given the name of
the product as an input. The script can be used by the AI as a tool to
locate the path to the mk file for the given product name, and enable
soong-only build if the verification (soong_only_diff_test.py) passes.

Test: run manually
Bug: 427983604
Change-Id: I2a7b24d6c3fcf69f99e453387229a87812a302c5