Fix `git_main-plus-rust` build errors where path != version
The branch `git_main-plus-rust` now uses a `.test` suffix for trying out
new toolchain builds. The existing Soong logic was relying on parsing
the path of the Rust toolchain as its version number as well. I split
the functionality now such that we have GetRustPrebuiltsVersion() and
GetRustVersion() which can be used to get either the full path to the
Rust prebuilt toolchain to test or the version number, respectively.

Bug: 421269002
Test: toolchain/android_rust/tools/dispatch-linux.sh main-plus-rust aosp_cf_x86_64_phone-trunk_staging-eng
Flag: EXEMPT refactor
Change-Id: I901d15329f1800640101fe49a5827e0a32bdc5c9