Add real deps on fingerprint file where used
There was some concern about using a real dependency on the fingerprint
file because it contains the build number, so sometimes an order-only
dependency was used. But the build number is now constant in local
builds, so a real dependency should not cause excessive rebuilds.

The order-only dependency was causing stale values in incremental
builds, as seen in b/421945305.

In make, `m installclean` is used to force a lot of this stuff to
rebuild. In soong-only, we don't want to rely on `m installclean`, so
instead we add real dependencies everywhere. The fact that the build
number only changes on CI will keep local build times fast while still
doing the necessary rebuilding on CI.

Bug: 421945305
Test: m nothing --no-skip-soong-tests
Change-Id: I03a258dbbe389a3755730fd32baa672c3c0ad279