Don't export transitive static lib header jars with combined implementation jars
In some cases when turbine is not supported the implementation jar
is used as the header jar.  If the implementation jar already has
its dependencies merged in, for example when jarjar is enabled,
then using it as the local header jar while combining it with
transitive static lib header jars results in duplicates of the
dependencies with and without jarjar applied.

Keep track of when the implementation jar contains dependencies,
and don't export the transitive static lib header jars when it
is used as a header jar.

Bug: 432416369
Test: java_test.go
Change-Id: Ib97394b3f049fea5dbf9f300e61806e0e320a757