Remove future symbolfile tags
This tag was not allowlisted in 8d5b9d0 ("validate tags for typos"), so
it has been effectively disabled for a long while. Based on discussion
in ag/34777071, remove this from map files processing and docs.

A new "draft" type will be added later for public APIs under
development, aliased to platform-only initially and with the intention
to be exposed to CTS eventually.

Bug: 434593830
Test: atest test_symbolfile
Change-Id: I58573bf4870cbc3cc0475e8fd2278074c15bae14