Refactor radio.go into a standalone bootloader.go
The current implementation of radio.go contains the build rules to
unpack radio.img and bootloader.img for AVB and disting. Some products
like tangorpro have bootloader.img, but not rqdio.img. To support
building these devices in soong-only mode, introduce a custom bootloader
module type

Details
- Create bootloader.go and copy bootloader specific code from radio.go
  into this new file.
- Update android_device's existing bootloader property into a non-path
  dependency
- Update android_device's target_files.zip generation. For products that
  use a bootloader that cannot be OTA updated (like cuttlefish), the
bootloader will be copied to IMAGES/. For products that use a bootloader
that can be OTA updated (like pixel), the bootlaoder will be copied to
RADIO/. This matches the make implementation.

Test: diff test for tangorpro
Bug: 424803505
Change-Id: I09d0eaaee19cc8f01c2f208aa217c1a24407536c