Rework ErrorRule and remove ErrorRuleNoQuotes
The double quotes in error_rule still evaluate backticks, even when
they're single-quote escaped inside the double quotes. A more reliable
way to print an error message is to always echo the result of
proptools.NinjaAndShellEscape(), without any further quoting. Rework
ErrorRule into a function that does that for the user, and unexport
the raw errorRule object.

Bug: 434265259
Test: m nothing --no-skip-soong-tests
Change-Id: Ib60c130674e0eec2a5fca38839f826d617e13015