Fix the source abi report depending on the abi diff
In ag/33834910, I made the binary depend on the abi diff reports
via a validation action instead of using the make-based
LOCAL_ADDITIONAL_DEPENDENCIES. However, the source abi report depends
on the final binary, so that change made the source report depend
on the diff test as well. This was a problem when running
create_reference_dumps.py, which builds the source abi reports and
copies them to the source tree. It's expected that you run this when
the diff test starts complaining, to update the golden that you're
diffing against. So when create_reference_dumps.py tried to build a
new golden, the diff test failed.

To break this dependency, copy the final output binary to a new location
and only do the diffs when doing that copy. Now both the source report
and the abi diff depend on the final binary at its first location, but
only the final binary at the second location depends on the abi diffs.

Bug: 419663728
Bug: 434265259
Test: Manually. Downloaded patchset 8 of ag/34689214, `m libutils`, then ran the create_reference_dumps.py command that the build failure told me to run.
Change-Id: Id0664128fb4581e1aa8df3e14b2512e3c73fa4d5