Remove ro.build.product from system/build.prop
ro.build.product is obsolete and can be replaced with
ro.product.device. It is removed from the build.prop file, but still
set from the init process with the value of ro.product.device for the
compatibility.

To read the build info from the product/etc/build.prop, add the build
info to product build prop. Init will read the most product-specific
properties prior to system properties. So duplicated props in
system/build.prop will be ignored.

Bug: 435546264
Test: getprop ro.build.product
Change-Id: I4b1c01872c16eb983dc05580d6f8c24f7d1abbda