Fix source/prebuilt apex selection discrepancy in soong only
prebuilt apexes are expected to be disabled in asan and hwasan. For make
packaging, this is done by calling HideFromMake in
GenerateAndroidBuildActions. This happens very late in the mutator
phase, which makes this special case unavailable to soong packaging.

This CL moves this special case PrebuiltSelectModuleMutator (postDeps)
to make this special case available to soong's android_filesystem

Bug: 436353330
Test: lunch lynx_hwasan-trunk-userdebug
Test: m nothing and verified that source `com.google.pixel.camera.hal`
is selected

Change-Id: I6f931f13bec1f914a8ed30648448436d02eacb0c