Introduce module context method ModulePhonyFiles(...)
Paths registered via CheckbuildFiles(...) are no longer built when
running `m <module name>`, and there are no easy way to register an
module artifact as a dependency of the module name phony target without
registering the artifact as an output file, which exposes the artifact
to other modules.

ModulePhonyFiles(...) is a solution to this problem, in that it
allows the module artifacts to be added as dependencies of the module
name phony target without exposing them to other modules for
consumption but only builds the artifacts for the users.

This change also removes the unused `defaultDeps` variable.

Test: m nothing
Bug: 427397928
Change-Id: Ia349c4edd5fbbf207f7b61b46f783cf82002640f