Revert changes for go/compatible-build-fingerprint
The effort was abandoned long ago. Note that by the time when the effort
was made, the change was in build/make. Those changes are gone because
system properties are now built in Soong. This change is in effect a
revert of the build/make changes.

Bug: 186786987
Bug: 188824341
Bug: 204222311
Reason: go/compatible-build-fingerprint is abandoned
Test: build
Flag: EXEMPT revert
Change-Id: I66eab02a10b06d7a3047f67c07db4386c46cef10