Extract prebuilt kernel path from PRODUCT_COPY_FILES
... instead of TARGET_KERNEL_PATH. Some products specify only
TARGET_KERNEL_DIR instead of TARGET_KERNEL_PATH, in which case fsgen
will fail to find the path to the kernel and will not create the boot
image.

Instead of exporting TARGET_KERNEL_DIR to Soong and replicating the
logic in Soong, this change utilizes the fact that the path to the
prebuilt kernel is specified as a PRODUCT_COPY_FILES entry if either of
TARGET_KERNEL_DIR or TARGET_KERNEL_PATH are set.

Test: build/soong/scripts/soong_only_diff_test.py tokay_experimental
Bug: 436905316
Change-Id: I6ea16e45b9158b9ca5487731c3f4f3dcec251810