Upgrade unused-variable warning from -Wno to -Wno-error
unused-variable is an important warning that helps catching issues. We
are in the effort of cleaning up all the issues in the platform, upgrade
the warning from silent to warning only while we finish off the
remaining instances. Since most the warnings have been cleaned up, the
amount of additional build log spam is small.

Test: presubmit
Bug: 315246135
Change-Id: I30fa1927a8cc8841003bd1336e9153ef0632113b