Pass kotlinc friend paths through the module xml file
The classpath is passed to kotlinc through the module xml file,
but the friend paths were being passed through the -Xfriend-path
argument on the command line.  The paths in the module xml file
are absolute (because if they were relative they would be interpreted
relative to the module xml file, not the top of the source tree),
but the paths on the command line have to be relative to avoid encoding
the absolute path to the source tree onto the command line.

Kotlin attempts to match up the relative paths in the friends path
and the absolute path in the classpath by calling toRealPath on the
relative path to make it absolute, which works most of the time.
This fails on CoG though because the out directory is a symlink, and
calling toRealPath on a path with a symlink resolves the symlink to
its target.  toRealPath is not called on the absolute path, so the
paths don't match, and the friend path is ignored.

Move the friend paths into the module xml file, where they can be
identical absolute paths to the classpath, and avoid needing the
complicated resolutions in kotlinc.

Bug: 423637257
Test: m SettingsRoboTests on CoG
Test: TestKotlinAssociates
Change-Id: I2d7f838790f8dc3816d9097495107bcaa8290109