Enable basic list iteration rewriting for R8 *and* D8
This is the default for R8, but not D8. Explicitly enable the
safe and straightforward optimization for both.

Note that, for pure D8 usage, this also requires using `--release`
mode, which is not yet the default due to stripping of debug info.
In practice, this change will have limited impact, but sets the default
for future optimizations as we broaden D8 release usage.

Bug: 379702549
Flag: EXEMPT basic bytecode optimization
Test: presubmit
Change-Id: I02f20afc5cd891adfe32d7f9db6facca3b3f88ac