Dist build.log file when build fails
This was broken when we made the diff test run two builds,
it exited early if the first one failed, causing it to not
get to the move_artifacts_to_subfolder() line and the build.log
not being available.

Also if a build fails the target files comparison report
won't be available, so don't attempt to open it to avoid
a crash.

Test: Manually
Change-Id: Ic7aa0b0a1cc2971cd9eaf41da424972175024569