Support install in root with PRODUCT_COPY_FILES in fsgen
PRODUCT_COPY_FILES may include entries like the followings:
`<src dir>:root/<dest dir>`
In such cases, the prebuilt_root module in system partition with
`install_in_root: true` property set should be generated and added as a
dep as the auto generated system image, as currently in fsgen a
dedicated root partition is not generated, but are included as part of
the system image.

This change adds support for such PRODUCT_COPY_ENTRIES instead of
omitting them.

Test: build/soong/scripts/soong_only_diff_test.py tangorpro_car
Bug: 438548759
Change-Id: I24965fc2d82b8eaf9d0aa638d3d674f028b16ac3