Convert sdk package to ModuleProxy
sdk.GenerateAndroidBuildActions was collecting dependency info in
collectMembers in the OSType-specific variants, and then retrieving
it from the CommonOS variant.  collectMembers is already using
WalkDepsProxy, change it to be run in the CommonOS variant and then
walk into the OSType-specific variants one at a time, collecting
the necessary info directly into the CommonOS variant where it will
be used.

Bug: 377723687
Test: sdk_test.go
Change-Id: I8002e649d1ca4949ee6a64c337c9e3c0c2e2d72a