Install unstripped boot images in symbols.zip
ag/35088302 restored the OAT files of boot images in the staging symbols
directory, but did not create the necesssary build rules to include
those in the device specific symbols.zip file. This CL adds those build
rules. Only the primary boot image will be included, but this can be
expanded to art and mainline boot images if necessary. Host variants
will not be included.

Bug: 438829484
Test: lunch husky_trunk-staging-userdebug
Test: SOONG_ONLY=true m out/soong/.intermediates/build/soong/fsgen/husky_generated_device/android_arm64_armv8-5a_cortex-a55/symbols.zip
Test: SOONG_ONLY=false out/target/product/husky/husky-symbols.zip
Test: verified that both symbols.zip contains out/target/product/husky/symbols/system/framework/arm64/boot.oat

Change-Id: I4c4553598c6dcde657e714ec50449e193cefa29d