Move libclang_rt.builtins for host linux_glibc out of sanitize
aosp/1241186 added a dependency on libclang_rt.builtins that I think was
supposed to only affect modules that depended on a static library that
needed the ubsan runtime or a minimal runtime, but instead affected any
module with a static library dependency, which is basically all of them.
That means we've been mixing a manual libclang_rt.builtins dependency
with the -lgcc -lgcc_s that the clang driver normally adds.

The dependency is now required as the hermetic libgcc in the tree is
too old and clang can emit references to symbols that are not provided
by libgcc but are available in libclang_rt.builtins.

Move the dependency out of the sanitizers and into linker.go where it
is clear that it applies to everything.

Eventually the explicit dependency can be removed when we switch to
passing --rtlib=compiler-rt and clang handles the dependency itself.

Bug: 440132952
Test: builds
Change-Id: If2d098c32c3c203ba113cffef1e0df048bfa1977