Implement update-api in soong_ui
We want to make the source tree read-only during the build, but
update-api writes to the source tree. Rewrite it so that it just builds
an update_api.txt file that contains information about all the files
that it wants to copy, and then soong_ui will read that file and do
the copies after the build is over and the read-only source tree sandbox
has been exited.

Bug: 200803952
Test: m update-api
Change-Id: I2e45274ae96761e9848f5ad673c658dda702c6b7