Add draft symbol tag for public APIs in development
Based on the discussion in ag/34777071, this add a draft symbol type for
public API symbols that are under development and not finalized.
Initially, this will be aliased to platform-only, and CTS access is only
possible with dlsym(). But a separate alias is introduced so that the
intention is clear, and if we figure out a solution with CTS and NDK
stubs such usage can be replaced with a dlsym()-less solution.

Bug: 434593830
Test: atest test_symbolfile
Change-Id: If54b620cf3e8b67a841d4b197495a7d6838028e3