Write .top file using defer
This write was moved to the bottom of the function in ag/35015964 to
avoid calling it if we run into an error during the function body, which
makes sense. However, bug reports with malformed symlinks have
increased, which means we are relying on this behavior in some edge
case. This cl mitigates the bug so we can find the proper fix.

Bug: 437350052
Change-Id: Idf00231465147402bc2f5e4bc5a37c26330efd6e