Use --treat-as-error instead of --warnings-as-errors
Previously, `--warnings-as-errors` would only treat `WARNING`s as
`ERRORS`. It would not affect `WARNING_ERROR_WHEN_NEW` as that would
defeat the whole purpose of it. However, the behavior of
`--warnings-as-errors` has been changed upstream so it now affects
`WARNING_ERROR_WHEN_NEW` as well. A new `--treat-as-error` option has
been added to allow more control over which severities are treated as
`ERROR`.

This change replaces the use of `--warnings-as-errors` with
`--treat-as-error warning` to replicate the previous behavior of
`--warnings-as-errors`.

Bug: 438246528
Test: m checkapi
Change-Id: I257b7358c4f393d910beeb490662561e1707b2cb