Fix incorrect handling of transitive required deps
When a module is overridden, the current logic in fsgen does not
correctly handle the required dependencies of the overridden module.
It installs the required dependencies even if no other module that is
being installed actually needs them.

For example, with the following setup:
    - bar requires foo
    - baz overrides bar
    - qux requires foo

If PRODUCT_PACKAGES contains bar baz, foo should not be installed.
If PRODUCT_PACKAGES contains bar baz qux, foo should be installed.

The current logic installs foo in both cases.

This change fixes the issue by tracking the reverse dependencies of
all modules. A required dependency of an overridden module is only
removed if all of its reverse dependencies are also overridden.

Bug: 439912806
Test: build/soong/scripts/soong_only_diff_test.py tangorpro_car
Change-Id: I04fca0020b37a34a115e51ca1fd57b7fb0f71e78