Fix incorrect single quote in compdb
If the flags contains quotes, the generated command would include
single-quote around the argument, which leads to incorrect commands
being generated in compdb, thus dropped by clangd from using it.

This commit trims the leading and trailing single quote from argument so
that the generated compdb is correct for clangd.

Bug: 442496243
Test: manual
Change-Id: Ib23bc09bff2c706d04c5188993f35ce8ecceecaa