Add phony targets to build host shared lib symlinks in testcases
This is particularly for BWYN to install host shared lib symlinks in
testcases folder per module.

Confirmed with presubmit test results, the zip is almost the same as the
current logic done in python script, but with a smaller size due to the
zip only includes the needed shared libs.

Bug: 437123882
Test: presubmit
Change-Id: I119f660f4ec36aef0e8db483a97a1aa0fccc7aa5