Fix data race in metrics events
The metrics events were written to expect a stack of nested events.
That expectation was violated when various background tasks were added
like starting RBE.

Replace the metrics.Metrics.EventTracer.Begin() method with
metrics.Metrics.Begin() that returns an event object that can be ended
directly by the caller instead of putting the event in a stack and
ending the top event when End is called.  Also add a lock to protect
the Metrics fields that can be set by multiple goroutines.

The metrics.Event is created by calls to Context.BeginTrace, which
now must also return a TraceEvent object that can be ended by the
caller.

Commands call AddProcResInfo and SetFatalOrPanicMessage, which were
previously adding the process info to the top event in the stack.  Pass
the TraceEvent to Commands so that they can call AddProcResInfo or
SetFatalOrPanicMessage directly on the correct event.

execution_metrics called BeginTrace and EndTrace, and must now call
End on the TraceEvent, but execution_metrics cannot import the build
package because the build package imports the execution_metrics
package.  To avoid the circular dependency define interfaces in
execution_metrics and implement them in build to adapt the TraceEvent
type into the execution_metrics package.

Bug: 308886397
Test: builds
Flag: EXEMPT HOST_MODE
Change-Id: Id9ffd9cda0b227c583e4af234eac7932d51783a8