Update APEX build to strip away beta flags from aconfig_flags.pb
For each mainline module, there is a unique beta namespace. The flags in
this namespace should not appear in aconfig_flags.pb but should be in
the storage file which we will use for API testing for B+ devices later.

This build system change is guarded by a build flag
RELEASE_REMOVE_BETA_FLAGS_FROM_ACONFIG_FLAGS_PB

Bug: 442853137
Test: tested with both turning the build flag on and off
Change-Id: Icd09f52cbd74086fe9075c083c0421f0d6e7e608