Support stripping debug symbols in ramdisk_16k module
When generating the ramdisk_16k image, the debugging symbols are
stripped for the prebuilt kernel modules that are not part of the GKI
entries before being copied to the staging directory in make.

This change adds support for the debug symbols stripping in Soong, by
introduing `strip_symbol_srcs` property in the ramdisk_16k module type.
Each entries listed in `strip_symbol_srcs` must be listed in `srcs` as
well, as the exact order in `srcs` are used to generate the modules.load
file if `load` property is not specified.

Test: build/soong/scripts/soong_only_diff_test.py aosp_cf_arm64_phone_pgagnostic
Bug: 442911131
Change-Id: Id3d10ab18b8b138ffaa9bbaeced7189b2ef4db86