Fix binary diffs in $ANDROID_PRODUCT_OUT/vbmeta_system.img
Soong's vbmeta module does a sort on the partitions to be added as
`--include_descriptors_from_image`. The sort order has been carefully
curated to match the make implementation.

This seems to be only necessary for the "primary" vbmeta.img file, and
not the secondary vbmeta_system.img and vbmeta_vendor.img files. Those
seem to respect the order specified BOARD_AVB_VBMETA_(SYSTEM|VENDOR).

This CL restricts the sort to the primary vbmeta.img file

Test: build/soong/scripts/soong_only_diff_test.py husky
Bug: 435530838
Change-Id: I66d456830fa92d39f4e8d8e345ecca7c2f3debfc