Support using @android.annotation.Hide instead of @hide
Previously, the `android.annotation.Hide` annotation was not supported
for all uses of `droidstubs`. It was only supported for the
`system-server` surface. That meant it was only possible to use an
`@hide` javadoc tag to hide source items from the API. This changes
that by adding `--hide-annotation android.annotation.Hide` to all
invocations of Metalava that consume sources and removes the no longer
required special handling for the `system-server` surface.

Flag: EXEMPT build refactoring
Bug: 444151311
Test: m checkapi
Change-Id: Ibd804028b678a4c2629034f3404212a647e1c68e