Introduce a .cache-marker file.
The .cache-marker file is removed before compilation begins and is
restored after it completes.

The absense of a .cache-marker file at the start of execution implies
the cache is invalid, and that non-incremental compilation should be
used. Any existing cache or output directories are first removed.

The .cache-marker file is not restored in the case of an OOM or
internal compiler error. We assume the cache is corrupted in those
cases. The marker _is_ restored in the event of a standard compilation
error, as this is assumed to be more of a syntactical mistake on the
part of the developer. The compiler should handle such cases
gracefully.

Fixes: 444187553
Test: various runs messing with .cache-marker file
Change-Id: I95b62bbf7d8abbea1f28d550ffcf83d83ff06d64